Separate BeatmapPlaylistItem into another file
