+++ osu.Game.Tests/Visual/TestCasePlaylist.cs
@@ -18,6 +18,7 @@ namespace osu.Game.Tests.Visual
         public override IReadOnlyList<Type> RequiredTypes => new[]
         {
             typeof(BeatmapPlaylist),
+            typeof(BeatmapPlaylistItem),
         };
 
         private BeatmapPlaylist playlist;
